mm-video: venc: Correct a typo in variable name
A kernel variable was to be defined as unsigned long but it is mistakenly
defined as unsigned int only, the space is missing after long. This bug is
silent because unsigned int is also a valid data type by itself.

Corresponding to kernel fix, similar correction is done in userspace code.

CRs-fixed: 556771
Change-Id: Ie58f275149dc9c85553f75e02594113b1a03ddcf
Signed-off-by: Deepak Verma <dverma@codeaurora.org>