mm-video-v4l2: venc: Protect buffer from being freed while accessing
Output buffer (in use-buffer mode) has an internal backup ION buffer. The
contents of this buffer are deep-copied in client's buffer in the context
of VideoEncCallBackThread; while this buffer can be freed in the client
thread's context.

Check the allocation bitmask before attempting to copy and synchronize
these operations by holding a lock.

Bug: 36130225
CRs-Fixed: 2053101
Fixes: Security Vulnerability - Heap use after free in libOmxVenc.
Change-Id: If5e89703b2dec0aee8acb7e897e9df94227af3f3
Signed-off-by: Praveen Chavan <pchavan@codeaurora.org>